message-admin-demo/pyproject.toml
[project]
name = "message-admin-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.49.1",
    "kafka-python>=2.2.15",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.12.0",
]
